Add Pundit authorization layer (finding #7)
- Add pundit gem
- ApplicationPolicy denies everything by default
- HouseholdPolicy, ActivityPolicy, ChildPolicy permit all actions for household members (flat membership model; roles can be layered in later)
- HouseholdMembersController authorizes via HouseholdPolicy#manage_members?
- verify_authorized after_action in ApplicationController ensures no action can accidentally skip authorization
- Devise controllers skip verify_authorized

+++ app/controllers/children_controller.rb
@@ -8,6 +8,7 @@ class ChildrenController < ApplicationController
   sig { void }
   def create
     child = @household.children.build(child_params)
+    authorize child
 
     if child.save
       redirect_to(household_path(@household), notice: "#{child.first_name} added.")
@@ -19,6 +20,7 @@ class ChildrenController < ApplicationController
   sig { void }
   def update
     child = @household.children.find(params[:id])
+    authorize child
 
     if child.update(child_params)
       redirect_to(household_path(@household), notice: "Updated.")
@@ -30,6 +32,7 @@ class ChildrenController < ApplicationController
   sig { void }
   def destroy
     child = @household.children.find(params[:id])
+    authorize child
     child.destroy
     redirect_to(household_path(@household), notice: "#{child.first_name} removed.")
   end
